% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltkPLUS-package.R
\docType{package}
\name{tcltkPLUS-package}
\alias{tcltkPLUS}
\alias{tcltkPLUS-package}
\title{tcltkPLUS: A tcltk Wrapper for R}
\description{
tcltkPLUS provides a collection of wrapper functions for the tcltk package in R. Functions are designed with ease of use in mind and aim to simplify GUI building in R. tcltkPLUS also provides an easy way for compiling R scripts into self-contained apps that can be run via .bat files.
}
\keyword{internal}
